markdown 解析代码分离，生成静态时复用 可复用
